masssive overhaul of internal documentation and structure
I've been doing some heavy housekeeping! I added a detailed intitial setup guide, added and improved commentary accross the whole program, and improved the general readability of the whole codebase. I also added checks for seeing if the official instance of the bot is active or not. This allows for other people to easily get their own branch of the bot running without dealing with all of my specialized features for the offical bot.