coin conversion command major rework
completely rewrote the coin conversion command to be more structured and easy to use. The data source for conversion rates is now privided by coin gecko so that the most coins possible are supported. In addition to these changes, help responses for incorrect usage have been added to improve user friendliness.