added checks for delisted pairs in stex function
Needed to check for missing data on a pair that DOES actually exist in the API. Missing data for an existing pair seems to be caused by that specific pairing recently having been removed, but still showing as available on the API side. To work around this, I've added a check that see's if there is price data after pulling a pair and then reporting and handling that situation accordingly. Along with this, I specified to have the stex client use the newer V2 API.