added duplicate ticker support to coin gecko command
This resolves the issue of there being multiple coins that share the same ticker. Now each coin will be displayed if multiple are seen under the same ticker. This applies to the coin gecko price command, as well as the conversion command because it uses coin gecko for its price data.