updated commands documentation
This was long overdue, I'm so sorry... I have now updated this doc to include the newest commands and also revise the usage of some of the older ones. This doc is now up to date with the current instance of the bot and will maintained to stay that way moving forward.