added error checks and fixed stats reporting
A pre-check is now done on all messages before attempting to process them as commands. This will cut down on the massive amounts of messages that come in where the channel they were sent from does not allow the bot to reply. This is mostly and issue with the spanish speaking servers where the prefixes clash with the language a bit.