added handling for coin gecko price calls on tickers that represent multiple coins
by default, the coin gecko quick price command will now show all instances of a ticker for cases where more than one coin exist with with the same ticker name.

This will not affect tbpa calls for the time being as this is still being sorted out right now.

Eventually the plan is to let users pick a personal preference of which coin to display in price and tbpa calls when there are multiple for a ticker. This implementation will at least let you see them all now until the personal preference stuff is ready!